requirements.txt
dataclass-csv==1.4.0
hurry.filesize==0.9
loguru==0.6.0
pydantic==1.10.2
PyYAML==6.0
typing-extensions==4.4.0
